requirements/base.txt
Django>=3.0,<3.1
django-dotenv==1.4.1
wagtail>=2.9,<2.10
wagtailfontawesome>=1.1.3,<1.2
Pillow>=6.2.0,<6.3
django-allauth==0.41.0
django-comments-xtd==2.6.1
